+++ Kernel/processQueue.c
@@ -156,15 +156,15 @@ void printQueue() { // FOR DEBUGGING
     node tmp=first;
     printString("PID    SLEEPS    NAME    ADDRESS    TAIL\n",255,255,255);
     for(i=0;i<queueSize;i++) {
-        printInt(tmp->process->pid,255,255,255);
+        printInt((uint64_t)tmp->process->pid,255,255,255);
         printString("       ",255,255,255);
-        printInt(tmp->process->sleeps,255,255,255);
+        printInt((uint64_t)tmp->process->sleeps,255,255,255);
         printString("      ",255,255,255);
         printString(tmp->process->name,255,255,255);
         printString("      ",255,255,255);
-        printInt(tmp,255,255,255);
+        printInt((uint64_t)tmp,255,255,255);
         printString("      ",255,255,255);
-        printInt(tmp->tail,255,255,255);
+        printInt((uint64_t)(tmp->tail),255,255,255);
         printString("\n",255,255,255);
 
         tmp = tmp->tail;
@@ -201,7 +201,8 @@ void removeLast()
 }
 
 
-int removeByPid(pid_t pid){
+int removeByPid(pid_t pid)
+{
     node tmp = first;
     node prev = NULL;
